update unstable 3.11.31 http://sourceforge.net/projects/math-atlas/files/Developer%20%28unstable%29/3.11.31/

+++ src/auxil/ATL_ptflushcache.c
@@ -1,5 +1,5 @@
 /*
- *             Automatically Tuned Linear Algebra Software v3.10.2
+ *             Automatically Tuned Linear Algebra Software v3.11.31
  *                    (C) Copyright 2000 R. Clint Whaley
  *
  * Redistribution and use in source and binary forms, with or without
@@ -37,7 +37,7 @@ struct ATL_FC
 {
    double dret;
    double *dp;
-   int N;
+   size_t N;
 };
